v1.9

+++ test/regress/pgbackrest_build.bash
@@ -2,6 +2,11 @@
 
 cd "$(dirname "$0")"
 
+BRANCH="master"
+if [ ! -z "$1" ]; then
+    BRANCH="$1"
+fi
+
 if [ -e /usr/bin/pgbackrest ]; then
 	INITIAL_VERSION=`/usr/bin/pgbackrest version | sed -e s/pgBackRest\ //`
 	echo "Initial pgBackRest version is : $INITIAL_VERSION"
@@ -13,19 +18,26 @@ if [ -e /usr/bin/pgbackrest ]; then
 	mv /usr/bin/pgbackrest /usr/bin/pgbackrest-$INITIAL_VERSION
 fi
 
-yum install -y gcc make openssl-devel libxml2-devel postgresql-devel perl-ExtUtils-Embed
+yum install -y make gcc postgresql-devel openssl-devel libxml2-devel lz4-devel libzstd-devel bzip2-devel
+
 
 if [ ! -d /build ]; then
 	mkdir /build
 else
-	rm -rf /build
+	rm -rf /build/pgbackrest
 fi
 
-yum install -y git
-git clone https://github.com/pgbackrest/pgbackrest.git /build/pgbackrest
+if [ "$BRANCH" == "local" ] && [ -e /pgbackrest-dev ]; then
+	echo "Build local pgbackrest-dev environment"
+	ln -s /pgbackrest-dev /build/pgbackrest
+else
+	yum install -y git
+	echo "Branch to clone is : $BRANCH"
+	git clone --single-branch --branch $BRANCH https://github.com/pgbackrest/pgbackrest.git /build/pgbackrest
+fi
+
 cd /build/pgbackrest/src && ./configure
 make -s -C /build/pgbackrest/src
-
 MAKE_VERSION=`/build/pgbackrest/src/pgbackrest version | sed -e s/pgBackRest\ //`
 echo "pgBackRest master version is : $MAKE_VERSION"
 mv /build/pgbackrest/src/pgbackrest /usr/bin/pgbackrest
